refactor(tests): split test_tsconfig.py into focused modules
Replace the 563-line monolithic test_tsconfig.py with four focused files:
- test_tsconfig_parse.py — parse_tsconfig (6 tests)
- test_tsconfig_extends.py — resolve_extends_chain (6 tests)
- test_tsconfig_discover.py — discover_tsconfigs (8 tests)
- test_tsconfig_partition.py — partition_files, has_explicit_scoping,
  create_temp_tsconfig (12 tests, +1 for new exclude scoping case)

Shared write_tsconfig helper moved to tsconfig_helpers.py.
